requirements.txt
fastapi==0.109.2
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiofiles==23.2.1
jinja2==3.1.2
python-multipart==0.0.6
openai==1.5.0
google-api-python-client==2.114.0
youtube-transcript-api==0.6.1
pytube==15.0.0
pdfdocument==3.3.0
markdown==3.5.2
beautifulsoup4==4.12.2
lxml==4.9.3 